Replacing DATETIME against TIME in timer element and rejecting any date part of data
There is no good reason to work with date format 'Y-m-d H:i:s' or '%Y-%m-%d %H:%i:%s'
as the value shown in form&details/list is a time 'H:i:s' anyway. 
Appearance of the value as datetime in database (e.g. '1999-01-01 00:02:34') is nothing else but confusing. See http://fabrikar.com/forums/showthread.php?t=30490




The plugin works fine on the TIME field. But 
NOTICE - this commit doesn't resolve another (not very serious) issue: 
In some unknown reason - when creating a new element the new databasefield is (and was) TEXT instead of TIME (earlier - instead of DATETIME). Fortunately the element seems to be working also well on this datatype, but this is rather confusing as well.