Also for Joomla3/F3.1 - autocreate pk element as internalid
This fix is already committed for F3.0, but 3.1 needs it as well. Without this the pk element was autocreated as "field" instead of "internalid" when new list was created on existing table. 
The same happened also to joined groups when new table joins were created via list admin to table without any list (and internalid element).